+++ packages/avatars-avataaars-sprites/src/utils/getOption.ts
@@ -1,4 +1,4 @@
-import type Options from '../options';
+import type { Options } from '../options';
 
 export default function (key: keyof Options, value: string, options: Options): boolean {
   let mode = options.mode || 'include';