+++ packages/avatars-human-sprites/src/index.ts
@@ -1,10 +1,7 @@
 import Male from '@dicebear/avatars-male-sprites';
 import Female from '@dicebear/avatars-female-sprites';
 import Random from '@dicebear/avatars/lib/random';
-
-type Options = {
-  mood?: Array<'happy' | 'sad' | 'surprised'>;
-};
+import type { Options } from './options';
 
 export function create(random: Random, options: Options = {}) {
   if (random.bool(50)) {