+++ packages/avatars-initials-sprites/src/options.ts
@@ -5,11 +5,10 @@
  * and run json-schema-to-typescript to regenerate this file.
  */
 
-export type ColorLevel = 50 | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
-export type Chance = number;
-
 export interface Options {
-  colors?: (
+  margin?: number;
+  background?: string | "transparent";
+  backgroundColors?: (
     | "amber"
     | "blue"
     | "blueGrey"
@@ -30,11 +29,8 @@ export interface Options {
     | "teal"
     | "yellow"
   )[];
-  colorful?: boolean | number;
-  primaryColorLevel?: ColorLevel;
-  secondaryColorLevel?: ColorLevel;
-  textureChance?: Chance;
-  mouthChance?: Chance;
-  sidesChance?: Chance;
-  topChance?: Chance;
+  backgroundColorLevel?: 50 | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
+  fontSize?: number;
+  chars?: number;
+  bold?: boolean | number;
 }